get simple binary ops working
